2016-2017/frameworks/requirements.txt
Flask == 0.11.*
Flask-SQLAlchemy == 2.1.*
Flask-WTF == 0.13.*
